Module aptos_transaction_builder::aptos_stdlib

Enums

ScriptFunctionCall

Structured representation of a call into a known Move script function.

Functions

encode_account_create_account

Basic account creation methods.

encode_account_rotate_authentication_key

encode_account_utils_create_and_fund_account

encode_reconfiguration_force_reconfigure

Force an epoch change.

encode_stake_add_stake

Add amount of coins from the account owning the StakePool.

encode_stake_increase_lockup

Similar to increase_lockup_with_cap but will use ownership capability from the signing account.

encode_stake_register_validator_candidate

Initialize the validator account and give ownership to the signing account.

encode_stake_rotate_consensus_key

Rotate the consensus key of the validator, it’ll take effect in next epoch.

encode_stake_set_operator

Allows an owner to change the operator of the stake pool.

encode_stake_unlock

Similar to unlock_with_cap but will use ownership capability from the signing account.

encode_stake_withdraw

Withdraw from account’s inactive stake.

encode_test_coin_claim_mint_capability

Claim the delegated mint capability and destroy the delegated token.

encode_test_coin_delegate_mint_capability

Create delegated token for the address so the account could claim MintCapability later.

encode_test_coin_mint

Create new test coins and deposit them into dst_addr’s account.

encode_test_coin_transfer

encode_token_create_limited_collection_script

encode_token_create_limited_token_script

encode_token_create_unlimited_collection_script

encode_token_create_unlimited_token_script

encode_token_direct_transfer_script

encode_token_initialize_token_for_id

encode_token_initialize_token_script

encode_token_transfers_cancel_offer_script

encode_token_transfers_claim_script

encode_token_transfers_offer_script

encode_transaction_publishing_option_set_module_publishing_allowed

encode_validator_set_script_add_validator

encode_validator_set_script_create_validator_account

encode_validator_set_script_create_validator_operator_account

encode_validator_set_script_register_validator_config

encode_validator_set_script_remove_validator

encode_validator_set_script_set_validator_config_and_reconfigure

encode_validator_set_script_set_validator_operator

encode_version_set_version

Updates the major version to a larger version.

encode_vm_config_set_gas_constants